// src/app/model/business-account-registration.ts
export interface BusinessDetail{
  company: string;
  email: string;
  biztype: string;
  phone: string;
}

export interface BusinessLocation{
  state: string;
  lga: string;
  address: string;
}

export interface BusinessContact{
  firstname: string;
  lastname: string;
  gender: string;
  phone: string;
  dob: string;
}

export interface BusinessCred{
  password1: string;
  password2: string;
}